refactor styling to somewhat fix header
